isbnカラムをbigintに変更（桁が足りないため）

+++ db/migrate/20200702105729_create_books.rb
@@ -5,7 +5,7 @@ class CreateBooks < ActiveRecord::Migration[5.2]
       t.string :author
       t.string :publisher
       t.string :image
-      t.integer :isbn
+      t.bigint :isbn
       t.string :url
       t.timestamps
     end
